Removed old HorizontalColorbar component
